refactored devices

+++ src/ch/hsr/ifs/liquids/util/Vector.java
@@ -8,11 +8,11 @@ public class Vector {
 		this(0, 0, 0);
 	}
 
-	public Vector(int x, int y) {
+	public Vector(float x, float y) {
 		this(x, y, 0);
 	}
 
-	public Vector(int x, int y, int z) {
+	public Vector(float x, float y, float z) {
 		this.x = x;
 		this.y = y;
 		this.z = z;
